refactor: 패키지 구조를 헥사고날 아키텍처에 맞게 재구성 및 도메인 영역과 S3 관련한 클래스 간에 의존성 제거를 위한 코드 변경
