+++ src/main/java/com/yoger/productserviceorganization/product/adapters/web/dto/response/SellableProductResponseDTO.java
@@ -1,4 +1,4 @@
-package com.yoger.productserviceorganization.product.dto.response;
+package com.yoger.productserviceorganization.product.adapters.web.dto.response;
 
 import com.yoger.productserviceorganization.product.domain.model.PriceByQuantity;
 import com.yoger.productserviceorganization.product.domain.model.Product;
