+++ src/main/java/com/yoger/productserviceorganization/product/adapters/web/dto/validation/ImageFileValidator.java
@@ -1,4 +1,4 @@
-package com.yoger.productserviceorganization.product.validation;
+package com.yoger.productserviceorganization.product.adapters.web.dto.validation;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
