+++ src/main/java/com/yoger/productserviceorganization/product/application/ProductService.java
@@ -1,14 +1,15 @@
-package com.yoger.productserviceorganization.product.domain;
+package com.yoger.productserviceorganization.product.application;
 
+import com.yoger.productserviceorganization.product.domain.port.ProductRepository;
 import com.yoger.productserviceorganization.product.domain.model.Product;
 import com.yoger.productserviceorganization.product.domain.model.ProductState;
-import com.yoger.productserviceorganization.product.dto.request.DemoProductRequestDTO;
-import com.yoger.productserviceorganization.product.dto.response.DemoProductResponseDTO;
-import com.yoger.productserviceorganization.product.dto.response.SellableProductResponseDTO;
-import com.yoger.productserviceorganization.product.dto.response.SimpleDemoProductResponseDTO;
-import com.yoger.productserviceorganization.product.dto.response.SimpleSellableProductResponseDTO;
+import com.yoger.productserviceorganization.product.domain.port.ImageStorageService;
+import com.yoger.productserviceorganization.product.adapters.web.dto.request.DemoProductRequestDTO;
+import com.yoger.productserviceorganization.product.adapters.web.dto.response.DemoProductResponseDTO;
+import com.yoger.productserviceorganization.product.adapters.web.dto.response.SellableProductResponseDTO;
+import com.yoger.productserviceorganization.product.adapters.web.dto.response.SimpleDemoProductResponseDTO;
+import com.yoger.productserviceorganization.product.adapters.web.dto.response.SimpleSellableProductResponseDTO;
 import com.yoger.productserviceorganization.product.mapper.ProductMapper;
-import com.yoger.productserviceorganization.product.persistence.ProductRepository;
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -20,7 +21,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class ProductService {
     private final ProductRepository productRepository;
-    private final S3Service s3Service;
+    private final ImageStorageService imageStorageService;
 
     public List<SimpleSellableProductResponseDTO> findSimpleSellableProducts() {
         return productRepository.findByState(ProductState.SELLABLE).stream()
@@ -34,8 +35,8 @@ public class ProductService {
 
     @Transactional
     public DemoProductResponseDTO saveDemoProduct(@Valid DemoProductRequestDTO demoProductRequestDTO) {
-        String imageUrl = s3Service.uploadImage(demoProductRequestDTO.image());
-        String thumbnailImageUrl = s3Service.uploadImage(demoProductRequestDTO.thumbnailImage());
+        String imageUrl = imageStorageService.uploadImage(demoProductRequestDTO.image());
+        String thumbnailImageUrl = imageStorageService.uploadImage(demoProductRequestDTO.thumbnailImage());
         Product product = ProductMapper.toDomainFrom(demoProductRequestDTO, imageUrl, thumbnailImageUrl);
         return DemoProductResponseDTO.from(productRepository.save(product));
     }
