+++ src/main/java/com/yoger/productserviceorganization/product/domain/port/ImageStorageService.java
@@ -1,0 +1,7 @@
+package com.yoger.productserviceorganization.product.domain.port;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface ImageStorageService {
+    String uploadImage(MultipartFile image);
+}
